add view past medication entries page
